# shared/openapi.yaml
openapi: 3.1.0
info:
  title: 'Web-block openapi'
  license:
    name: MIT
    identifier: MIT
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/'
    description: 'Backend server'
paths:
...
  /api/user:
    get:
      tags:
        - user
      operationId: 971db9bc974c0e607288cfe7ecc10096
      responses:
        '200':
          description: 'User data.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
    put:
      tags:
        - user
      operationId: fd4f39f7820d3f3cbc259c5b30fed830
      requestBody:
        description: 'User data to update.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '204':
          description: 'User updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
    delete:
      tags:
        - user
      operationId: db0c8ea1c902ed078a87cf1fa6497994
      requestBody:
        description: 'User password'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteUserRequest'
      responses:
        '204':
          description: 'User deleted.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
    patch:
      tags:
        - user
      operationId: ea88adacb890b864b59bcce5aff2473d
      requestBody:
        description: 'User data to update.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '204':
          description: 'User updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
  /api/user/style:
    put:
      tags:
        - user
      operationId: f8741720bfaafb5c8126e8c1418a9917
      requestBody:
        description: 'User style data to update.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStyleRequest'
      responses:
        '204':
          description: 'User style updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
    patch:
      tags:
        - user
      operationId: 54ba995c0c13dd114980733de45f7ffa
      requestBody:
        description: 'User style data to update.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStyleRequest'
      responses:
        '204':
          description: 'User style updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
  /api/user/password:
    put:
      tags:
        - user
      operationId: 57ddbc5bebe0cd30c268a0d8a4640660
      requestBody:
        description: 'User password to update.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '204':
          description: 'User password updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
    patch:
      tags:
        - user
      operationId: 75f8f589d8162c3ebb0a27049232e804
      requestBody:
        description: 'User password to update.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '204':
          description: 'User password updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
  /api/user/repos:
    get:
      tags:
        - user
      operationId: 395814df2599dc7a564535035ede73fb
      responses:
        '200':
          description: 'User repos from github.'
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: number
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
  '/api/user/repo/{name}':
    get:
      tags:
        - user
      operationId: 5dcb8b09ef61f0dd409ac3fbf1c59d74
      parameters:
        -
          name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Repos url.'
          content:
            application/json:
              schema:
                type:
                  - string
                  - 'null'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
components:
  schemas:
    ErrorObject:
      required:
        - message
        - errors
      properties:
        message:
          type: string
          example: 'Validation failed.'
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
              example: 'Field is required'
      type: object
...
    PanelPosition:
      type: string
      enum:
        - top-left
        - top-right
        - top-center
        - bottom-left
        - bottom-center
        - bottom-right
    Themes:
      type: string
      enum:
        - light
        - dark
        - cupcake
        - bumblebee
        - emerald
        - corporate
        - synthwave
        - retro
        - cyberpunk
        - valentine
        - halloween
        - garden
        - forest
        - aqua
        - lofi
        - pastel
        - fantasy
        - wireframe
        - black
        - luxury
        - dracula
        - cmyk
        - autumn
        - business
        - acid
        - lemonade
        - night
        - coffee
        - winter
        - dim
        - nord
        - sunset
...
    DeleteUserRequest:
      required:
        - password
      properties:
        password:
          type: string
          format: password
      type: object
...
    UpdatePasswordRequest:
      description: 'Request to update user password'
      required:
        - oldpassword
        - newpassword
        - newPassword_confirmation
      properties:
        oldPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
        newPassword_confirmation:
          type: string
          format: password
      type: object
...
    UpdateStyleRequest:
      $ref: '#/components/schemas/Style'
      description: 'Request for updating user style'
    UpdateUserRequest:
      description: 'Request to update user information'
      required:
        - name
        - email
        - password
        - settings
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type:
            - string
            - 'null'
          format: password
        settings:
          required:
            - lang
            - hideExtensions
            - defaultBranch
          properties:
            lang:
              type: string
              format: enum
              enum:
                - en
                - lv
            hideExtensions:
              type: boolean
            defaultBranch:
              type: string
              default: main
          type: object
      type: object
...
    Keybinds:
      description: 'User keybinds'
      required:
        - quickCommand
      properties:
        quickCommand:
          type: string
          default: Control+p
      type: object
    Style:
      description: 'User style'
      required:
        - controlPosition
        - minimapPosition
        - pathPosition
        - baseLightTheme
        - baseDarkTheme
      properties:
        controlPosition:
          allOf:
            -
              $ref: '#/components/schemas/PanelPosition'
        minimapPosition:
          $ref: '#/components/schemas/PanelPosition'
        pathPosition:
          $ref: '#/components/schemas/PanelPosition'
        baseLightTheme:
          $ref: '#/components/schemas/Themes'
        baseDarkTheme:
          $ref: '#/components/schemas/Themes'
      type: object
    Settings:
      description: 'User settings.'
      required:
        - hideExtensions
        - defaultBranch
        - lang
        - style
        - keybinds
      properties:
        hideExtensions:
          type: boolean
          default: true
        defaultBranch:
          type: string
          default: master
        lang:
          type: string
          default: en
        style:
          $ref: '#/components/schemas/Style'
        keybinds:
          $ref: '#/components/schemas/Keybinds'
      type: object
    User:
      required:
        - name
        - email
        - settings
      properties:
        name:
          type: string
        email:
          type: string
        has_github:
          type: boolean
        is_admin:
          type: boolean
        settings:
          $ref: '#/components/schemas/Settings'
      type: object
  securitySchemes:
    sessionAuth:
      type: apiKey
      description: "This API uses session-based authentication. The session cookie is automatically handled by the \n            middleware, and clients do not need to include it in their requests."
      name: web_block_session
      in: cookie
      flows:
        implicit:
          authorizationUrl: /api/auth/redirect
          scopes:
            - read
            - write
...